feat: separate scallop pool from scallop
